rename controllers to avoid duplicate names
