+++ app/Http/Controllers/Api/Manager/ManagerUserController.php
@@ -1,30 +1,42 @@
 <?php
 
-namespace App\Http\Controllers\Api\Admin;
+namespace App\Http\Controllers\Api\Manager;
 
-use App\Actions\Admin\User\UserCreateAction;
-use App\Actions\Admin\User\UserUpdateAction;
+use App\Actions\Manager\User\UserCreateAction;
+use App\Actions\Manager\User\UserDisassociateFromOrganizationAction;
+use App\Actions\Manager\User\UserUpdateAction;
 use App\DTO\Shared\UserDTO;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Api\Admin\UserCreateRequest;
-use App\Http\Requests\Api\Admin\UserUpdateRequest;
-use App\Http\Resources\Api\Admin\UserResource;
+use App\Http\Requests\Api\Manager\UserCreateRequest;
+use App\Http\Requests\Api\Manager\UserUpdateRequest;
 use App\Http\Resources\Api\Shared\PaginationResource;
 use App\Http\Resources\Api\Shared\UserListWithRolesResource;
+use App\Models\Organization;
 use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
-class UserController extends Controller
+class ManagerUserController extends Controller
 {
     /**
      * @OA\Get(
-     * path="/api/admin/users",
-     * operationId="AdminGetUsers",
-     * tags={"Admin/User"},
+     * path="/api/manager/users/{organization}",
+     * operationId="ManagerGetUsers",
+     * tags={"Manager/User"},
      * summary="Get a list of users",
      * description="Retrieve a list of users.",
      * security={{"sanctum":{}}},
+     *
+     *     @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="The organization id",
+     *         required=true,
+     *
+     *         @OA\Schema(
+     *             type="integer"
+     *         )
+     *     ),
      *
      *     @OA\Parameter(
      *         name="q",
@@ -74,13 +86,21 @@ class UserController extends Controller
      *     ),
      * )
      */
-    public function index(): JsonResponse
+    public function index(Organization $organization): JsonResponse
     {
         try {
-            $this->authorize('view', User::class);
+            $this->authorize('viewByOrganization', [User::class, $organization]);
 
-            $search = request()->get('q', null);
-            $paginate = User::search($search)->paginate(30);
+            $query = User::query()
+                ->whereHas('organizations', function ($query) use ($organization) {
+                    return $query->where('organization_id', $organization->id);
+                });
+
+            if ($search = request()->get('q', null)) {
+                $query->whereRaw("LOWER(name) LIKE '%' || LOWER(?) || '%'", [$search]);
+            }
+
+            $paginate = $query->paginate(30);
 
             return response()->json([
                 'type' => 'success',
@@ -95,12 +115,23 @@ class UserController extends Controller
 
     /**
      * @OA\Post(
-     *     path="/api/admin/users",
-     *     operationId="AdminCreateUser",
-     *     tags={"Admin/User"},
+     *     path="/api/manager/users/{organization}",
+     *     operationId="ManagerCreateUser",
+     *     tags={"Manager/User"},
      *     summary="Create a new user",
-     *     description="Create a new admin user with the provided information.",
+     *     description="Create a new user with the provided information in manager.",
      *     security={{"sanctum":{}}},
+     *
+     *     @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="The organization id",
+     *         required=true,
+     *
+     *         @OA\Schema(
+     *             type="integer"
+     *         )
+     *     ),
      *
      *     @OA\RequestBody(
      *         required=true,
@@ -156,15 +187,15 @@ class UserController extends Controller
      *     ),
      * )
      */
-    public function store(UserCreateRequest $request): JsonResponse
+    public function store(UserCreateRequest $request, Organization $organization): JsonResponse
     {
         try {
-            $this->authorize('create', User::class);
+            $this->authorize('createByOrganization', [User::class, $organization]);
 
             $data = $request->validated();
 
             $userDto = new UserDTO($data);
-            UserCreateAction::execute($userDto);
+            UserCreateAction::execute($userDto, $organization);
 
             return response()->json(['type' => 'success', 'message' => __('messages.common.success_create')], HttpResponse::HTTP_OK);
         } catch (\Exception $e) {
@@ -174,9 +205,9 @@ class UserController extends Controller
 
     /**
      * @OA\Put(
-     *     path="/api/admin/users/{user}",
-     *     operationId="AdminUpdateUser",
-     *     tags={"Admin/User"},
+     *     path="/api/manager/users/{user}",
+     *     operationId="ManagerUpdateUser",
+     *     tags={"Manager/User"},
      *     summary="Update user",
      *     description="Update user with the provided information.",
      *     security={{"sanctum":{}}},
@@ -264,12 +295,23 @@ class UserController extends Controller
 
     /**
      * @OA\Delete(
-     *     path="/api/admin/users/{user}",
-     *     operationId="AdminDestroyUser",
-     *     tags={"Admin/User"},
+     *     path="/api/manager/users/{user}/{organization}",
+     *     operationId="ManagerDestroyUser",
+     *     tags={"Manager/User"},
      *     summary="Destroy user",
      *     description="Destroy user with the provided id.",
      *     security={{"sanctum":{}}},
+     *
+     *     @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="The organization id",
+     *         required=true,
+     *
+     *         @OA\Schema(
+     *             type="integer"
+     *         )
+     *     ),
      *
      *     @OA\Parameter(
      *         name="id",
@@ -305,12 +347,12 @@ class UserController extends Controller
      *     ),
      * )
      */
-    public function destroy(User $user): JsonResponse
+    public function disassociateUserFromOrganization(User $user, Organization $organization): JsonResponse
     {
         try {
-            $this->authorize('delete', $user);
+            $this->authorize('disassociateUserFromOrganization', $user);
 
-            $user->delete();
+            UserDisassociateFromOrganizationAction::execute($user, $organization);
 
             return response()->json(['type' => 'success', 'message' => __('messages.common.success_destroy')], HttpResponse::HTTP_OK);
         } catch (\Exception $e) {
@@ -320,9 +362,9 @@ class UserController extends Controller
 
     /**
      * @OA\Get(
-     *     path="/api/admin/users/{user}",
-     *     operationId="AdminGetUser",
-     *     tags={"Admin/User"},
+     *     path="/api/manager/users/{user}",
+     *     operationId="ManagerGetUser",
+     *     tags={"Manager/User"},
      *     summary="Get user infos",
      *     description="Get user infos",
      *     security={{"sanctum":{}}},
@@ -369,10 +411,12 @@ class UserController extends Controller
     public function getUser(User $user): JsonResponse
     {
         try {
+            $this->authorize('viewByUserOrganizations', $user);
+
             return response()->json([
                 'type' => 'success',
                 'message' => __('messages.common.success_view'),
-                'data' => UserResource::make($user),
+                'data' => UserListWithRolesResource::make($user),
             ], HttpResponse::HTTP_OK);
         } catch (\Exception $e) {
             return response()->json(['type' => 'error', 'message' => $e->getMessage()], HttpResponse::HTTP_BAD_REQUEST);
