+++ tests/Feature/Controllers/Api/Manager/ManagerUserControllerTest.php
@@ -11,7 +11,7 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 use Tests\TestCase;
 
-class UserControllerTest extends TestCase
+class ManagerUserControllerTest extends TestCase
 {
     use RefreshDatabase;
